Restructured geometry helper functions
