+++ src/geom.js
@@ -1,0 +1,6 @@
+var getcenter = require('./geometry/getcenter');
+
+var geom = {};
+geom.center = getcenter;
+
+module.exports = geom;
