+++ src/maputils.js
@@ -64,34 +64,5 @@ module.exports = {
       else {
         return undefined;
       }
-  },
-  getCenter: function getCenter(geometry) {
-      var type = geometry.getType();
-      var center;
-      switch (type) {
-          case "Polygon":
-              center = geometry.getInteriorPoint().getCoordinates();
-              console.log(center);
-              break;
-          case "MultiPolygon":
-              center = geometry.getInteriorPoints()[0].getCoordinates();
-              break;
-          case "Point":
-              center = geometry.getCoordinates();
-              break;
-          case "MultiPoint":
-              center = geometry[0].getCoordinates();
-              break;
-          case "LineString":
-              center = geometry.getCoordinateAt(0.5);
-              break;
-          case "MultiLineString":
-              center = geometry.getLineStrings()[0].getCoordinateAt(0.5);
-              break;
-          case "Circle":
-              center = geometry.getCenter();
-              break;
-      }
-      return center;
   }
 }
